Always use a new connection for websockets
Don't try to reuse an existing connection when servicing a request to upgrade a connection to websockets.

Websockets connections can't be reused. An attempt to reuse a websockets connection for HTTP will fail.

Fixes an issue where a websockets connection is attempted to be reused resulting in garbage and failures when that connection is reused.